+++ rich/console.py
@@ -26,6 +26,7 @@ from typing import (
     Mapping,
     NamedTuple,
     Optional,
+    Set,
     TextIO,
     Tuple,
     Type,
@@ -325,14 +326,17 @@ class Capture:
 
     Args:
         console (Console): A console instance to capture output.
+        echo (bool): True if captured content should also be written to the
+            terminal. False otherwise.
     """
 
-    def __init__(self, console: "Console") -> None:
+    def __init__(self, console: "Console", echo: bool = False) -> None:
         self._console = console
+        self._echo = echo
         self._result: Optional[str] = None
 
     def __enter__(self) -> "Capture":
-        self._console.begin_capture()
+        self._console.begin_capture(self._echo)
         return self
 
     def __exit__(
@@ -551,6 +555,7 @@ class ConsoleThreadLocals(threading.local):
     theme_stack: ThemeStack
     buffer: List[Segment] = field(default_factory=list)
     buffer_index: int = 0
+    buffer_indices_to_render: Set[int] = field(default_factory=set)
 
 
 class RenderHook(ABC):
@@ -744,8 +749,12 @@ class Console:
         )
 
         self._record_buffer_lock = threading.RLock()
+
+        buffer_index = 0
         self._thread_locals = ConsoleThreadLocals(
-            theme_stack=ThemeStack(themes.DEFAULT if theme is None else theme)
+            theme_stack=ThemeStack(themes.DEFAULT if theme is None else theme),
+            buffer_index=buffer_index,
+            buffer_indices_to_render={buffer_index},
         )
         self._record_buffer: List[Segment] = []
         self._render_hooks: List[RenderHook] = []
@@ -782,6 +791,10 @@ class Console:
     @_buffer_index.setter
     def _buffer_index(self, value: int) -> None:
         self._thread_locals.buffer_index = value
+
+    @property
+    def _buffer_indices_to_render(self) -> Set[int]:
+        return self._thread_locals.buffer_indices_to_render
 
     @property
     def _theme_stack(self) -> ThemeStack:
@@ -863,9 +876,16 @@ class Console:
         """Exit buffer context."""
         self._exit_buffer()
 
-    def begin_capture(self) -> None:
-        """Begin capturing console output. Call :meth:`end_capture` to exit capture mode and return output."""
+    def begin_capture(self, echo: bool = False) -> None:
+        """Begin capturing console output. Call :meth:`end_capture` to exit capture mode and return output.
+
+        Args:
+            echo (bool): True if captured content should also be written to the
+                terminal. False otherwise.
+        """
         self._enter_buffer()
+        if echo:
+            self._buffer_indices_to_render.add(self._buffer_index)
 
     def end_capture(self) -> str:
         """End capture mode and return captured string.
@@ -875,6 +895,7 @@ class Console:
         """
         render_result = self._render_buffer(self._buffer)
         del self._buffer[:]
+        self._buffer_indices_to_render.discard(self._buffer_index)
         self._exit_buffer()
         return render_result
 
@@ -1084,9 +1105,13 @@ class Console:
         """Play a 'bell' sound (if supported by the terminal)."""
         self.control(Control.bell())
 
-    def capture(self) -> Capture:
+    def capture(self, echo: bool = False) -> Capture:
         """A context manager to *capture* the result of print() or log() in a string,
         rather than writing it to the console.
+
+        Args:
+            echo (bool): True if the captured output should also be written to the console,
+                False otherwise.
 
         Example:
             >>> from rich.console import Console
@@ -1098,7 +1123,7 @@ class Console:
         Returns:
             Capture: Context manager with disables writing to the terminal.
         """
-        capture = Capture(self)
+        capture = Capture(self, echo=echo)
         return capture
 
     def pager(
@@ -1986,7 +2011,8 @@ class Console:
                 buffer_extend(line)
 
     def _check_buffer(self) -> None:
-        """Check if the buffer may be rendered. Render it if it can (e.g. Console.quiet is False)
+        """Render the buffer if possible. Clear it if required.
+        The buffer may not be rendered in the event that Console.quiet is True.
         Rendering is supported on Windows, Unix and Jupyter environments. For
         legacy Windows consoles, the win32 API is called directly.
         This method will also record what it renders if recording is enabled via Console.record.
@@ -1999,13 +2025,12 @@ class Console:
                 with self._record_buffer_lock:
                     self._record_buffer.extend(self._buffer[:])
 
-            if self._buffer_index == 0:
+            if self._buffer_index in self._buffer_indices_to_render:
 
                 if self.is_jupyter:  # pragma: no cover
                     from .jupyter import display
 
                     display(self._buffer, self._render_buffer(self._buffer[:]))
-                    del self._buffer[:]
                 else:
                     if WINDOWS:
                         use_legacy_windows_render = False
@@ -2063,6 +2088,11 @@ class Console:
                             raise
 
                     self.file.flush()
+
+                is_capture = self._buffer_index != 0
+                if not is_capture:
+                    # The Capture context still needs access to the buffer to read the captured content,
+                    # so we can't clear it here. It's the responsibility of the Capture context to clear it.
                     del self._buffer[:]
 
     def _render_buffer(self, buffer: Iterable[Segment]) -> str: