+++ tests/test_console.py
@@ -384,7 +384,7 @@ def test_capture_and_record(capsys):
     recorder = Console(record=True)
     recorder.print("ABC")
 
-    with recorder.capture() as capture:
+    with recorder.capture(echo=False) as capture:
         recorder.print("Hello")
 
     assert capture.get() == "Hello\n"
@@ -395,6 +395,33 @@ def test_capture_and_record(capsys):
     assert recorded_text == "ABC\nHello\n"
     assert capture.get() == "Hello\n"
     assert out == "ABC\n"
+
+
+def test_capture_echo_outputs_captured_content_to_terminal(capsys):
+    console = Console()
+
+    console.print(1)
+    with console.capture(echo=True) as capture:
+        console.print(2)
+    console.print(3)
+
+    out, err = capsys.readouterr()
+    assert capture.get() == "2\n"
+    assert out == "1\n2\n3\n"
+
+
+def test_capture_echo_nested_capture():
+    # TODO: This behaviour doesn't seem correct
+    console = Console()
+    console.print(1)
+    with console.capture() as capture1:
+        console.print(2)
+        with console.capture() as capture2:
+            console.print(3)
+    console.print(4)
+
+    assert capture1.get() == ""
+    assert capture2.get() == "2\n3\n"
 
 
 def test_input(monkeypatch, capsys):